Added history locking.
History locking is now available by right-clicking a menu item. A bold
label will indicate a locked record that will not be cleared by history.

This commit also fixes several bugs around the relationship between
clipboard, daemon, and provider, notably when the daemon would seem to
preempt the control of the clipboard.

The GUI now buffers the history entry items, as opposed simply the text.
This allows for finer-grained control, in the future, over item
selection should it become necessary.

An unfortunate consequence of this change set, however, sees additional
coupling between GUI and history through the clipboard. As this seems
rather complex, it will be a subject of focus.