Proper handling for NULL clipboard values.
When the clipboard receives a NULL value, the first, unlocked entry in
the history is removed. This allows external applications, like key
managers, to properly clear the clipboard.

This does come with some drawbacks, such as unintentional history
removal if and when a NULL value is expected to be written. Indeed, this
is currently deemed a worthwhile sacrifice, as it conforms with the
principle of least surprise.

Further, when a window that owned the clipboard was closed, it may
have set the clipboard to NULL. As such, when the NULL was read by Clip,
it was deemed a request to remove the head history elment.

To resolve this issue, this changeset also introduces ownership
tracking, which will analyze the first clipboard value after an
ownership transfer. Iff the first value is a NULL, the former value is
restored.

This changeset also introduces a significant bugfix that resolves a data
race when waiting for GTK to return the clipboard data contents. This
should prevent Clip from seeing log artefacts when changing clipboard
values.