Feature, "disabling history" disables clipboard.
Previously, disabling the history would only disable new values from
being appended to the history. While this satisfied the feature's name,
it didn't do it in a meaningful way.

In such a state, clip still hijacked the actual clipboard content, which
is unnecessary and annoying.  This change allows for better
interoperation with programs that use the clipboard for non-textual
content.