Add tag support and refactor history.
This commit introduces basic (database/model) support for tagging and
cleans up a bunch of historical oddities. Numerous queries and functions
have been removed.

As an unfortunate result of this, a database change is required.
Specifically, the 'locked' column becomes a non-null integer and a new
column, 'tag' is added. The best way to apply these changes is to drop
(or rename) your history table (migration is up to you).