Add marking (like VIm).
Clip now supports the ability to tag/mark an entry for quick selection.
To mark a selected record, press 'm' and then the desired mark letter.
For example, pressing 'ma' will mark the record with the tag, 'a'.

To activate a marked record, press the gave character, '`', followed by
the desired mark. For example, to select the previously-selected 'a'
record, press '`a'. If the mark is found, the correlated record will be
activated.