Change how modifiers interact with the clipboard.
To prevent partial selections to be copied into PRIMARY, Clip previously
assumed that depressing the Shift key or left mouse button was intent
that the current value should not yet be copied. This worked well,
however, it interfered with marking when an upper-case mark was used.

The new behaviour allows consumers to specify whether or not the
clipboard value change is optional, as per the traditional rules, or
mandatory.