Add ability to mask entries.
Entries may be masked, on screen, by pressing '*'. The selected value
will be displayed as a string of asterisks, rendering the actual value
hidden.

Importantly, masking does not encrypt or obfuscate the actual value
stored in Clip's database. Masking exists simply to assist in mitigating
(but not eliminating) the risk of shoulder surfing.